Refactor createPlot.R code
I ditched ggghost and concatenated the code as text using paste and glue as in ggplotui, boxplotR or PlotsOfData. While ggghost was convenient, this allows more freedom in how the final code is shown to the user, which is an important factor of this app.
The old createPlot.R has been renamed to createPlotLegacy.R and will be finally deleted in the following commits when testing is finished.